django/requirements.txt
Django==4.2
djangorestframework
markdown
django-filter
#cx_Oracle
djangorestframework-simplejwt
pyyaml
django-cors-headers
mysqlclient